+++ skbio/stats/tests/test_composition.py
@@ -17,12 +17,13 @@ import scipy
 import copy
 from skbio import TreeNode
 from skbio.util import assert_data_frame_almost_equal
-from skbio.stats.composition import (
-    closure, multiplicative_replacement,
-    perturb, perturb_inv, power, inner,
-    clr, clr_inv, ilr, ilr_inv, alr, alr_inv,
-    sbp_basis, _gram_schmidt_basis, tree_basis,
-    centralize, _holm_bonferroni, ancom)
+from skbio.stats.distance import DistanceMatrixError
+from skbio.stats.composition import (closure, multiplicative_replacement,
+                                     perturb, perturb_inv, power, inner,
+                                     clr, clr_inv, ilr, ilr_inv, alr, alr_inv,
+                                     sbp_basis, _gram_schmidt_basis,
+                                     centralize, _holm_bonferroni, ancom,
+                                     vlr, pairwise_vlr, tree_basis)
 
 from scipy.sparse import coo_matrix
 
@@ -1231,6 +1232,77 @@ class AncomTests(TestCase):
         guessed_p = _holm_bonferroni(p)
         for a, b in zip(corrected_p, guessed_p):
             self.assertAlmostEqual(a, b)
+
+
+class TestVLR(TestCase):
+    def setUp(self):
+        self.mat = np.array([[1, 1, 2], [3, 5, 8], [13, 21, 55]])
+        self.mat_neg = np.array([[-1, 1, 2], [3, -5, 8], [13, 21, -55]])
+        self.mat_with_zero = np.array([[0, 1, 2], [3, 5, 8], [13, 21, 55]])
+
+    def test_vlr(self):
+        # No zeros
+        output = vlr(
+            x=self.mat[0],
+            y=self.mat[1],
+            ddof=1,
+            robust=False,
+        )
+        self.assertAlmostEqual(output, 0.0655828061998637)
+
+        # With zeros
+        output = vlr(
+            x=self.mat_with_zero[0],
+            y=self.mat_with_zero[1],
+            ddof=1,
+            robust=False,
+        )
+        assert np.isnan(output)
+
+        # assert raises error
+        with self.assertRaises(ValueError):
+            vlr(
+                x=self.mat_neg[0],
+                y=self.mat_neg[1],
+                ddof=1,
+                robust=False,
+            )
+
+    def test_robust_vlr(self):
+        # No zeros
+        output = vlr(
+            x=self.mat[0],
+            y=self.mat[1],
+            ddof=1,
+            robust=True,
+        )
+        self.assertAlmostEqual(output, 0.0655828061998637)
+
+        # With zeros
+        output = vlr(
+            x=self.mat_with_zero[0],
+            y=self.mat_with_zero[1],
+            ddof=1,
+            robust=True,
+        )
+        self.assertAlmostEqual(output, 0.024896522246558722)
+
+    def test_pairwise_vlr(self):
+
+        # No zeros
+        dism = pairwise_vlr(self.mat, ids=None, ddof=1, robust=False)
+        output = dism.condensed_form().sum()
+        self.assertAlmostEqual(output, 0.2857382286903922)
+
+        # With zeros
+        with self.assertRaises(DistanceMatrixError):
+            pairwise_vlr(self.mat_with_zero, ids=None, ddof=1, robust=False)
+
+        # no validation
+        dism = pairwise_vlr(self.mat, ids=None, ddof=1, robust=False,
+                            validate=False)
+        output = dism.data.sum() / 2
+        self.assertAlmostEqual(output, 0.2857382286903922)
 
 
 if __name__ == "__main__":